made hotel.controllers.js connected it to route index file
