+++ api/controllers/hotels.controllers.js
@@ -1,0 +1,6 @@
+module.exports.hotelsGetAll = function(res, res){
+     console.log("GET the JSON");
+        res
+            .status(200)
+            .json( {"jsonData" : true} );
+};
