+++ api/routes/index.js
@@ -1,0 +1,23 @@
+var express = require('express');
+var router = express.Router();
+
+/* 
+
+to link the hotels.controller to the route 
+
+*/
+var ctrlHotels = require('../controllers/hotels.controllers.js');
+
+/*
+
+Start adding routes to the router object by chaining it with the route  */
+/* You define your route then define your method then define your function you want to run ... you can chain different methods to different routes 
+
+*/
+router
+    .route('/hotels')
+    .get(ctrlHotels.hotelsGetAll);
+ 
+
+
+module.exports = router;
